pyproject.toml
[tool.poetry]
name = "splunk-connect-for-snmp-mib-server"
version = "0.1.0"
description = "Host a community collection of SNMP MIBs and serves them to SNMP Trap and SNMP Poller to do MIB translation"
authors = ["Your Name <you@example.com>"]
license = "APACHE-2.0"

[tool.poetry.dependencies]
python = "^3.8"
pymongo = "^3.11.3"
pysnmp = "^4.4.12"
requests = "^2.25.1"
PyYAML = "^5.4.1"
Flask = "^1.1.2"
Flask-AutoIndex = "^0.6.6"

[tool.poetry.dev-dependencies]
mongomock = "^3.22.1"

[tool.poetry.scripts]
sc4snmp-mib-server = "splunk_connect_for_snmp_mib_server.snmp_mib_server:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

